+++ src/bindings/utils.cpp
@@ -39,14 +39,4 @@ std::vector<double> invert(const std::vector<double> &vector) {
     return new_vector;
 }
 
-std::string replace_all(std::string str, const std::string &from,
-                        const std::string &to) {
-    size_t start_pos = 0;
-    while ((start_pos = str.find(from, start_pos)) != std::string::npos) {
-        str.replace(start_pos, from.length(), to);
-        start_pos += to.length();
-    }
-    return str;
-}
-
 }  // namespace vips