+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v0.0.5] - TBD
 
-Uses libvips v8.13.3, compiled with Emscripten v3.1.34.
+Uses libvips v8.14.2, compiled with Emscripten v3.1.34.
 
 ### Added
 
@@ -32,6 +32,7 @@ Uses libvips v8.13.3, compiled with Emscripten v3.1.34.
 
 - Use a stack size of 256kb for both main and newly created threads.
   Previously, this was configured at 5mb and 2mb, respectively.
+- Update methods/enums for libvips 8.14.
 
 ## [v0.0.4] - 2022-11-03